+++ src/components/Categories.jsx
@@ -1,15 +1,29 @@
-import React, { useContext } from 'react';
+import React, { useContext, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import Button from '../helpers/Button';
 import RecipesContext from '../contexts/RecipesContext';
+import {
+  getMealsRecipes,
+  getMealsByCategory,
+} from '../helpers/MealsAPI';
+import {
+  getCocktailsRecipes,
+  getCocktailsByCategory,
+} from '../helpers/CocktailsAPI';
 
 export default function Categories(props) {
   const { data } = props;
+
+  const maxCards = 12;
+
   const {
     selectedCategory,
+    setData,
     setSelectedCategory,
     setToggle,
     toggle,
+    type,
+    setIsFetching,
   } = useContext(RecipesContext);
 
   const filter = (strCategory) => {
@@ -23,6 +37,25 @@ export default function Categories(props) {
       setSelectedCategory('All');
     }
   };
+
+  useEffect(() => {
+    const filterCategory = async () => {
+      setIsFetching(true);
+      let results;
+      if (selectedCategory === 'All') {
+        results = (type === 'meal')
+          ? await getMealsRecipes() : await getCocktailsRecipes();
+      } else {
+        results = (type === 'meal') ? await getMealsByCategory(selectedCategory)
+          : await getCocktailsByCategory(selectedCategory);
+      }
+      if (results && selectedCategory) {
+        setData(results.filter((item, index) => index < maxCards));
+      }
+      setIsFetching(false);
+    };
+    filterCategory();
+  }, [selectedCategory, toggle]);
 
   return (
     <aside>
