+++ src/pages/MainPage.jsx
@@ -1,39 +1,14 @@
-import React, { useContext } from 'react';
-import MainCards from '../components/MainCards';
-import RecipesContext from '../contexts/RecipesContext';
-import Header from '../components/Header';
+import React from 'react';
 import Footer from '../components/Footer';
+import Header from '../components/Header';
 
-export default function MainPage() {
-  const { data, isFetching, type, categoriesData } = useContext(RecipesContext);
-  let title;
-  let thumbnail;
-  let strTitle;
-  let typeId;
-
-  if (type === 'meal') {
-    title = 'Comidas';
-    strTitle = 'strMeal';
-    thumbnail = 'strMealThumb';
-    typeId = 'idMeal';
-  } else {
-    title = 'Bebidas';
-    strTitle = 'strDrink';
-    thumbnail = 'strDrinkThumb';
-    typeId = 'idDrink';
-  }
-
-  return isFetching ? <p>Loading...</p> : (
-    <>
-      <Header title={ title } hasSearchBar />
-      <MainCards
-        data={ data }
-        thumbnail={ thumbnail }
-        title={ strTitle }
-        categories={ categoriesData }
-        typeId={ typeId }
-      />
+function MainPage() {
+  return (
+    <div>
+      <Header />
       <Footer />
-    </>
+    </div>
   );
 }
+
+export default MainPage;
